Send rating difference between player and opponent (UCI)
Send "UCI_RatingAdv" to communicate the rating difference between the
player and the opponent. The difference is computed as the rating
advantage. That is:

  UCI_RatingAdv = OwnRating - OpponentRating

UCI_RatingAdv can be used for contempt based on the relative opponent
strength, for instance.

UCI_RatingAdv is sent only when the ratings for both engines are
defined.